+++ api/webservice/Portal/BaseModule/SaveInventory.php
@@ -43,28 +43,12 @@ class SaveInventory extends \Api\Core\BaseAction
 		$recordModel = \Vtiger_Record_Model::getCleanInstance($moduleName);
 		$recordModel->set('subject', $moduleName . '/' . date('Y-m-d'));
 		$recordModel->initInventoryData(
-			$this->createClass($moduleName, $inventory)->getInventoryData(),
+			(new \Api\Portal\BaseModel\SaveInventory($moduleName, $inventory))->getInventoryData(),
 			false
 		);
 		$recordModel->save();
 		return [
 			'id' => $recordModel->getId(),
 		];
-	}
-
-	/**
-	 * Create class.
-	 *
-	 * @param string $moduleName
-	 *
-	 * @return \Api\Portal\BaseModel\AbstractSaveInventory
-	 */
-	protected function createClass(string $moduleName, array $inventory): \Api\Portal\BaseModel\AbstractSaveInventory
-	{
-		$className = "Api\\Portal\\{$moduleName}Model\\SaveInventory";
-		if (class_exists($className)) {
-			return new $className($moduleName, $inventory);
-		}
-		return new \Api\Portal\BaseModel\SaveInventory($moduleName, $inventory);
 	}
 }
